Fix #7 - require -RoboOnly for exclusions
Specifically, require -RoboOnly if you also specify -ExcludeDirectory or -ExcludeFile. People might think they work even with -ComOnly (but then the 6 additional properties are not displayed, but then again if you don't use either -ComOnly or -RoboOnly, COM is the default).

It makes sense to make people aware of this explicitly.